requirements.txt
matplotlib==3.9.2
numpy==1.26.4
#numpy==2.1.2
pandas==2.2.3
pillow==11.0.0
psycopg2==2.9.9
python-dotenv==1.0.1
scikit-learn==1.5.1
scipy==1.13.1
seaborn==0.13.2
tqdm==4.66.5
pipreqs==0.5.0


# torch
# torchaudio
# torchvision
# torchinfo
torch==2.5.0
torchaudio==2.5.0
torchinfo==1.8.0
torchmetrics==1.5.1
torchvision==0.20.0

pytorch-lightning==2.4.0
lightning==2.4.0
lightning-utilities==0.11.8